refactor(cli): extract session exit/signal subsystem from runChat
runChat was an ~870-line function holding ~14 co-captured closures for the
SIGINT/SIGTERM/SIGHUP/SIGTSTP/SIGCONT handlers, the double-tap-to-exit
confirmation, and the two teardown paths, with teardown scattered across the
body (mode-disable at three sites, signal wiring split between the closure and
the finally).

Lift that subsystem into createSessionExitController(ctx). Every closure body is
unchanged; the runtime session values they used to capture directly now arrive
through an explicit SessionExitControllerContext, and the graceful teardown that
was inline in the waitUntilExit finally becomes controller.gracefulTeardown().
runChat now creates the controller once Ink has mounted, calls install() to take
signal ownership, and its finally is a single gracefulTeardown() call; App and
the slash-command context reach the handlers through the controller.

Pure code motion, no behavior change. runChat drops ~243 lines and the exit
subsystem is one cohesive, documented unit instead of a closure tangle. The
signal/exit paths (Ctrl-C, Ctrl-Z/fg, SIGTERM) have no automated coverage, so
they warrant a manual smoke test.
